requirements.txt
python-dotenv
requests
smtplib
